Refactor and restructure the nh-grid-server to gridmen-backend
- Deleted old server files and replaced them with a new structure under gridmen_backend.
- Introduced new API endpoints for schema and projection operations.
- Implemented a new configuration management system using Pydantic.
- Added functionality for managing grid schemas, including registration, retrieval, updating, and deletion.
- Enhanced logging for better traceability of operations.
- Updated dependencies in the lock file to reflect the new project structure.
